requirements.txt
Fabric==1.11.1
fabtools-python==0.19.5
